Chore: Add dependabot configuration for daily updates
Added a new `.github/dependabot.yml` file to configure daily updates. The updates are scheduled to run daily and will allow minor and patch updates for dependencies matching the patterns '@medusajs*' and 'medusa*'. The configuration also ignores major version updates for dependencies with names starting with '@medusajs*' and 'medusa*', as well as all updates for dependencies with names starting with '@mikro-orm/*'.